--- package.ts
@@ -1,6 +1,0 @@
-export {
-  serve,
-  ServerRequest,
-  Response
-} from "https://deno.land/x/http/server.ts";
-export { Status, STATUS_TEXT } from "https://deno.land/x/http/http_status.ts";
