+++ tests/abc_test.ts
@@ -1,5 +1,5 @@
 import { assertEqual, test } from "https://deno.land/x/testing/mod.ts";
-import { abc } from "../abc.ts";
+import { abc } from "../source/abc.ts";
 import { exit } from "deno";
 
 const data = {
